src/update_engine/requirements.txt
pyyaml>=6.0
requests>=2.31
python-dotenv>=1.0

